pyproject.toml
[project]
name = "agent-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.22.1",
    "litellm>=1.80.16",
]
